[wpa] Remove PMKID checking
PMKID checking is an additional pre-check that helps detect invalid passphrases
before going through the full handshaking procedure. It takes up some amount of
code size, and is not necessary from a security perspective. It also is
implemented improperly by some routers, which was causing gPXE to give spurious
authentication errors. Remove it for these reasons.

Signed-off-by: Joshua Oreman <oremanj@rwcr.net>