[linker] Use weak definitions instead of weak declarations
This removes the need for inline safety wrappers, marginally
reducing the size penalty of weak functions, and works around an
apparent binutils bug that causes undefined weak symbols to not
actually be NULL when compiling with -fPIE (as EFI builds do).

A bug in versions of binutils prior to 2.16 (released in 2005)
will cause same-file weak definitions to not work with those
toolchains. Update the README to reflect our new dependency on
binutils >= 2.16.

Signed-off-by: Joshua Oreman <oremanj@rwcr.net>