[settings] Introduce TAG_TYPE().
TAG_TYPE() is a new macro to extract a type from the tag field of each
setting and from the tag_magic field of each settings block.  A Setting
and a settings block are compatible if and only if their TAG_TYPE()s
match.  TAG_TYPE() makes the compatibility testing independent of the
particular module, while maintaining setting numbering for backwards
compatibility.

For clarity and consistency:
  Make SMBIOS_TAG_MAGIC be 32 bits, the size of settings.tag_magic.
  Explicitly name DHCP_TAG_MAGIC, replacing 0's found in code.

TAG_READONLY() is included in this patch to show why TAG_TYPE() uses
"0xfd" instead of "0xff".  These definitions were chosen for backwards
compatibility with the previous tag_magic values, which were
	0x00000000 for DHCP settings
	0x5b000000 for SMBIOS settings
	0xc19c1900 for Phantom CLP settings.
TAG_READONLY() indicates which settings cannot be set by users and will
be used to clean up the settings_ui in a forthcoming patch.