[settings_ui] Highlight locally-set variables.
Display locally-set settings in bold, and leave settings defined in
children unhighlighted.  This way, you see the value of every setting
that is valid at the current scope, but values inheritted from child
settings blocks are relatively dim.

Also, display an explicit message when a setting is inheritted from a child.
This makes it easy to figure out why those entries are grayed.