[settings_ui] Add hierarchy navigation.
Add scope hierarchy navigation to the settings UI.  (Previously, the
settings UI could only edit settings of a fixed scope, specified on the
"config <scope>" command line.)

Add "parent" and "child" scope rows at the top of the scrollable listing
in the UI, which can be used to navigate to the parent and child scopes.

This patch abstracts the displayed "settings" as "row objects" and
adds types of row objects for parent and child scopes.  The bulk of the
patch is concerned with this substitution, and with the handling of the
new types.

Transitioning from one scope to the next was achieved as follows:
main_loop() was modified to return a pointer to the next settings block
that should be navigated, or NULL to exit; and its caller was modified
to re-call main_loop() to edit the next settings block.